Fix _resolve and improve inquire_reply
Inquire was returning the key instead of the value it pointed to in the options map. This was confusing until I realized its `_prepare` function was never called, which is where it supposed to be converting the result value.

I also improved the `_inquire_reply`, which was expecting the result to always be a boolean. So no matter what you entered as an option, if the value was not explicitly False it always printed out "Yes". Instead it checks that the value is a bool, and then prints 'Yes'/'No' accordingly, otherwise it prints out what ever the result itself is.