Inquire: Check value is a key and not a default
If the user hits enter, we will receive a default value instead of the key to a value. So lets check the dict for the result to see if it exists. If it does, returned the mapped value, otherwise we have a default value and should just return it as is.